+++ f4-renderer/include/f4/renderer/entity_render.hpp
@@ -61,7 +61,7 @@ namespace f4::renderer {
 class RenderResources;
 
 }  // namespace f4::renderer
-namespace f4::world_types { struct ClassTable; }
+namespace f4::world_types { class ClassTable; }
 namespace f4::renderer {
 
 // ---------------------------------------------------------------------------