+++ f4-renderer/include/f4/renderer/world_renderer.hpp
@@ -45,7 +45,7 @@
 #include <functional>
 #include <vector>
 
-namespace f4::world_types { struct ClassTable; }
+namespace f4::world_types { class ClassTable; }
 
 namespace f4::renderer {